Updating test harness

+++ source/TrekkingForCharity.Api.Write/Commands/CreateWaypointCommand.cs
@@ -17,7 +17,7 @@ namespace TrekkingForCharity.Api.Write.Commands
 
         public double Lat { get; set; }
 
-        public int WhenToReach { get; set; }
+        public long WhenToReach { get; set; }
 
         public Guid TrekId { get; set; }
     }
