+++ source/TrekkingForCharity.Api.Write/Commands/UpdateTrekCommand.cs
@@ -15,7 +15,7 @@ namespace TrekkingForCharity.Api.Write.Commands
 
         public string Description { get; set; }
 
-        public int WhenToStart { get; set; }
+        public long WhenToStart { get; set; }
 
         public string BannerImage { get; set; }
